refactor: 모달 및 별점 서비스 동작 정제 및 안정성 향상
